feat: refactor function engine
