+++ runtimes/nodejs/src/support/engine/index.ts
@@ -1,4 +1,4 @@
-export * from './engine'
+export * from './cache'
 
 export * from './function'
