+++ runtimes/nodejs/src/support/policy.ts
@@ -8,7 +8,7 @@
 import assert = require('assert')
 import { Params, Policy } from 'database-proxy'
 import { logger } from './logger'
-import { CloudFunction } from './function-engine'
+import { CloudFunction, FunctionCache } from './engine'
 import { DatabaseAgent } from '../db'
 import { generateUUID } from './utils'
 import { POLICY_COLLECTION } from '../constants'
@@ -95,12 +95,12 @@ export class PolicyAgent {
     }
 
     try {
-      const func_data = CloudFunction.getFunctionByName(injectorName)
-      assert.ok(func_data, 'getFunctionByName(): function not found')
+      const funcData = FunctionCache.get(injectorName)
+      assert.ok(funcData, 'getFunctionByName(): function not found')
 
-      const func = new CloudFunction(func_data)
-      const ret = await func.invoke({
-        __function_name: func.name,
+      const func = new CloudFunction(funcData)
+      const ret = await func.execute({
+        __function_name: funcData.name,
         requestId: generateUUID(),
       })
       assert(typeof ret.data === 'function', 'function type needed')
